Make permutation and compression steps NUMA aware. This might improve the permutation and compression steps with many threads. However, it might also be necessary to keep some of the big arrays used in the compression step local to the node. This I have not done yet.
I have extended run_threaded() with a parameter that takes either LOW or HIGH.
If LOW, a lower number of threads is used.
This lower number is specified with the -l option.

The number of threads specified with -t and -l is rounded up to a multiple
of the number of nodes.

Most invocations of run_threaded use HIGH.
In compress.c and compress_tmpl, I use high which is HIGH if NUMA and low otherwise.

If a particular invocation of run_threaded is problematic with many threads, then invoke with LOW instead of high/HIGH. I have not yet taken into account the indications in your PR.

rtbgen seems to generate correct output.