Support for generating Shatranj tables with 70-move rule.
Fairy-Stockfish uses a 70-move rule for Shatranj, and this seems to be
the generally accepted rule:
https://www.abstractgames.org/shatranj1.html

I have tested this on 4- and 5-piece tables, and the statistics seem to
be compatible with those generated with the 50-move rule. The only table
with cursed wins is KQBvKB.

This update also adds a -p option to the generator for ply-accurate DTZ
values.