TBPATH variable and bugfixes.
Instead of RTBWDIR (or JTBWDIR, etc.), the generator now searches for
tablebases in the directories listed in RTBPATH (or JTBPATH, etc.).
Directories are separated by ':' on Linux and ';' on Windows, as usual.

Various bugfixes to the Shatranj tablebase generator, which should now
hopefully be bug free.

To verify the correctness of generated Shatranj tablebases, jtbver and
jtbverp can be used. In addition, rtbver and rtbverp are now able to
verify files with very large DTZ values. The verification algorithm is
not guaranteed to catch all errors, but in practice it probably will
detect any remaining bugs that might affect a table.

I removed the -d option from the tbver/tbverp programs, which allowed
verification of tables in the (R)TBWDIR/TBZDIR directories. Instead,
one can now prepend the tablename with a path, e.g. "rtbver
../tb/KQvKR" will verify ../tb/KQvKR.rtbw and ../tb/KQvKR.rtbz.